feat(refactoring): migrate NgRx to domain-driven structure
