+++ src/app/components/tasks/tasks.component.spec.ts
@@ -3,7 +3,7 @@ import { provideMockStore } from '@ngrx/store/testing';
 import { NoopAnimationsModule } from '@angular/platform-browser/animations';
 
 import { TasksComponent } from './tasks.component';
-import { initialState } from '../../store/reducers/task.reducer';
+import { initialState } from '../../store/tasks/task.reducer';
 
 describe('TasksComponent', () => {
   let component: TasksComponent;
