+++ src/app/store/categories/index.ts
@@ -1,0 +1,4 @@
+export * from './category.actions';
+export * from './category.reducer';
+export * from './category.selectors';
+export * from './category.effects';
