+++ src/app/store/tasks/index.ts
@@ -1,0 +1,4 @@
+export * from './task.actions';
+export * from './task.reducer';
+export * from './task.selectors';
+export * from './task.effects';
