+++ src/app/store/tasks/task.selectors.ts
@@ -1,5 +1,5 @@
 import { createFeatureSelector, createSelector } from '@ngrx/store';
-import { TaskState } from '../reducers/task.reducer';
+import { TaskState } from './task.reducer';
 
 export const selectTaskState = createFeatureSelector<TaskState>('tasks');
